rename get_emoji module to emoji
